requirements.txt
Pillow
matplotlib
transformers
numpy
fastapi
uvicorn
pydantic
requests
python-multipart
opencv-python
torch 
torchvision 
torchaudio